app/script: make Appdash properly display traces for pages that have no AJAX
In the event that a page does not have any AJAX content on it at all (e.g. the homepage),
it is important that we still 'finalize' the PageLoad portion of the trace and display a
link to the user. To do this we simply add a premature timeout in case no ajaxStart/ajaxStop
events fire on the page, ensuring that everything works normally after the delay period.