app: remove extraneous invite links and DisableInvites/--app.disable-invites flag due to marginal value
Now that the homepage has a + button for adding users, the invite
links have less value. It also adds needless complexity that isn't
worth it IMO. Finally, given that the links went to the "AppURL and
DNS" doc page, it's unlikely that they would have served their purpose
(as that page isn't optimized for answering their question).