sgx: properly use the default endpoint in src-auth file.
Prior to this change [CS 22](https://src.sourcegraph.com/sourcegraph/.changesets/22)
broke the ability to use the default endpoint as in e.g.:

  $ src --endpoint=foobar.com login
  $ src repo create foo/bar
  ... oops it tries to create it on localhost:3000 ! ...

This is due to the fact that I switched to usage of EndpointURL (which defaults
to localhost:3000 in the event of no user endput for a --endpoint). This is not
what the code here wants, because it uses it's own default in the event of no
user input (the one specified in the auth file).

The behavior now matches identically with the behavior of it before CS 22 was
merged.