discussions: Use permalink as notif url
Before we used the DefKey. Since we can have multiple discussions per DefKey,
this is not as exact.

Test Plan: Created a discussion and commented on a discussion. Both times I
mentioned a person. I then inspected then clicked on the URLs in the logs for
the ignored slack messages.