server/internal/store/fs: improve Changesets.List performance significantly
Improve the performance of Changesets.List operations significantly by making
use of a single-syscall-operation to determine which changesets are open or
closed. Prior to this change, we would read in every single changeset metadata
file one-by-one. 170 of these added up to around 1.5s in total, and due to the
global FS lock this could cause requests to queue for much larger periods of
time.