notify: Forward GenericEvent to mothership
If email is not configured on an instance, we can forward our request to the
mothership to send the email for us. Currently we allow anyone to send these
emails, but stubs are in place to restrict this appropriately.

Test Plan: Have a working local mothership and local private instance. Configure
mothership with Mandrill secrets.

Update the discussions commentcreate to use the gRPC service instead
of the notif package. Create a comment. Observe logs to see instance forward to
demo-mothership. See demo-mothership send email. Read email in inbox :)