Add SSH git server.
It allows using the Git Transfer Protocol over SSH. It uses SSH public
keys
that users add to identify the Sourcegraph user for authorization
purposes.
Shell access is not enabled, but a message is printed as a way of
checking
that your username has been successfully authenticated.

Also change default SSH port to 3002.