platform: Make baseURI a URL
A platform app currently doesn't have a nice way to get absolute paths to
itself. `BaseURI` implies it would include this information, but instead it is
just the path component of the URL. A side effect of this is the URLs
for slack notifications are broken for changesets.

This will change the `BaseURI` from something like `/lib/annotate/.changes/1` to
`http://localhost:3000/lib/annotate/.changes/1`

Test Plan: Click around a bunch in the discussions and changesets app. Comment
on a changeset and check in logs that we use the correct url for the slack
notification

 Ignored Slack message: "*keegancsmith* reviewed <http://localhost:3000/lib/annotate/.changes/1|lib/annotate changeset #1>: Test /cc @keegancsmith\n\ntest".