Revert "platform: Make baseURI a URL"
This reverts commit 7a8f22b0d8f2bc15c9ca54c8ddffeb61752eb71b.

Reverting because there is a build failure in the tests (Test_repoFrameBaseURI).

I tried to fix the test, but it's not trivial and I don't have enough context
to be sure my fix would be correct. Reverting so I can apply my changes and
ensure tests are passing.

This commit should be updated so its tests build and pass, then reapplied.
/cc @keegancsmith