Address review comments.
Add a prometheus gauge for active SSH connections to git server.

Increase hard git transaction timeout to 30 minutes to allow for all
valid transactions to finish successfully. We can lower this time once
we gather data and see that it should be lowered.

Calculate the absolute path to repo, and check that it's inside the
repos root.