Don't checkout base branch during merge
Fixed an issue when trying to merge into a base branch that’s not
master. On the initial pull in NewRepoStage, the base branch is named
master (git’s default behavior), so if you try to merge branch foo into
branch bar, when you perform git checkout bar it fails because bar is
actually called master locally. This fixes that by removing the
unnecessary git checkout.