changesets: allow commenting on context-expanded lines
This change adds support for commenting on lines that are not a part of the diff (i.e.
the lines that appear when you click the top or bottom context expansion buttons).

From a UX perspective, we will also want to auto-expand such areas if someone did
comment on them (so they are easier to discover). This will be addressed as a separate
change.