sgx: properly determine cache key based on auth token
This fixes a critical issue which was the cause of mysterious behavior
on Sourcegraph. The grpc cache key on the client is determined using
the grpc endpoint and the credentials set in the context. However, for
federated requests, the auth token is not present in ctx credentials,
but in the ctx metadata. Before this commit, such requests would be
assumed to be unauthenticated requests and assigned the same cache key
even for different auth tokens in the metadata.

This is only a problem when we have a gRPC call with the same argument
values but different auth tokens in the ctx metadata. As it turns out,
this is exactly the case for Auth.Identify, which takes a pbtypes.Void
argument, and parses the actor from the ctx's auth token. Now, a
federated Auth.Identify request is made to the mothership for every
local gRPC call, to parse the auth token set in the context. So if this
endpoint's result is wrongly cached, it can affect the behavior of all
other gRPC endpoints.

Additionally, Auth.Identify sets a shortCache(ctx) cache trailer, which
instructs the gRPC client to cache the result for 60 seconds. This means
that when many users are hitting the local gRPC server, the actor set in
the context could be one of the actors from requests in the last 60
seconds. Hence we would get wrong names in the Slack notifications, as
well as UserKeys.Add would assign the key to a wrong user id.

The final piece of this puzzle is why would 'keegancsmith's login be set
in the ctx actor an abnormally high number of times, even when he wasn't
active on src.sourcegraph.com. The answer to this is that our repo is
mirrored on GitHub under his account, which would be doing background
polling with his credentials, so his authInfo would be most likely to be
cached as the Auth.Identify result at any point of time.