changeset: publish notifications from within gRPC methods
This will ensure that notifications are posted even when a changeset
is created/updated/merged from the CLI.

Additionally, this moves the JIRA update calls to the changeset listener,
so even JIRA integration will work when the changeset is modified from
the CLI.

Eventually, the changeset gRPC service must be moved to the platform
app package. But till then, the gRPC service cannot figure out the
URL to a changeset, so this is offloaded to the listener of changeset
events.