sshgit: use ctx authed with non self-signed tokens
cliCtx is authed with self-signed (HMAC) tokens which are not valid
on remote servers (eg for federated calls). This commit creates a
new ctx for sshgit server with a token scoped to internal:sshgit.

Additionally, done some refactoring in RegisteredClients methods:

- Previously in GetUserPermissions, only admin users on a client
  could view permission levels of a user on that client.
- Now any user with access to that client can view others'
  permissions, as long as the request originates from the same client.
- If the gRPC request is made from a different client, then the
  requesting user must be an admin on the client whose permissions
  are requested.

This reduces the constraint on GetUserPermissions in order to enable
the ssh server to internally verify a user's authorization levels.