Re-arrange initialization order in serve command.
This is needed so that the tracker app can register its frame (which
effectively adds web page handlers) before the web server is started.

It's also logically more sound and simple, as we no longer need to
handle the small window where the web server is started but some of the
background workers (that the app depends on) are not yet running.