Fix disappearing platform app tabs issue.
Fix problem in platform.Frames where ordered IDs included all apps,
even when not all apps are enabled for given repo.

This resulted in odd behavior for rendering app tabs (some apps
missing, etc.).

Improve documentation and consistency of the relevant code, and
simplify it where possible.

This is a followup to changes in
https://src.sourcegraph.com/sourcegraph/.changes/272. /cc  @beyang
@neelance