access: Save token from root auths
Once we have an access token for a fed root, we might as well save it and use it
for future login attempts. This gives a much better user experience upon logging
into a second sourcegraph instance.

```
  $ rm ~/.src-auth
  $ src --endpoint=http://localhost:3080 login
  Enter credentials for https://sourcegraph.com
  Username: keegancsmith
  Password:
  # Credentials for https://sourcegraph.com saved to /Users/keegan/.src-auth.
  # Credentials for http://localhost:3080 saved to /Users/keegan/.src-auth.
  # Default endpoint set to http://localhost:3080.
  $ src --endpoint=http://localhost:3080 login
  Using saved auth token for http://localhost:3080
  $ src --endpoint=https://src.sourcegraph.com login
  Using saved auth token for https://sourcegraph.com
  # Credentials for https://src.sourcegraph.com saved to /Users/keegan/.src-auth.
  # Default endpoint set to https://src.sourcegraph.com.
```

Notice how on attempting to relogin to `http://localhost:3080` we did nothing
since our old token was still valid. Also notice when logging into
`https://src.sourcegraph.com` we did not query for `https://sourcegraph.com`
username and password since we had already saved the auth token for it.