gitserver/sshgit: Parse and emit git events.
This change makes the ssh path emit the same git events as the https
path.

That way, they get sent as slack notifications, and changesets are
updated, etc.

It duplicates some code, but only 2 places for now, so refactoring felt
too soon. Should be done in the future as needed.