changesets: Use a lock-free next ID algorithm
This allows us to remove needing to hold the lock for creating a Changeset.
Every other operation relies on either Storage being safe for concurrency and is
namespaced to the current CS. This is true. The only other assumption we rely on
is that a branch still exists, but that is impossible to gaurentee even without
the lock + we return a valid error in such conditions.

We also rely on a CS being in index_all but not existing otherwise not breaking
anything else. From testing this is the case.