server/local: Avoid counting all commits in listCommitsUncached.
Optimize listCommitsUncached by calculating StreamResponse.HasNext
boolean value more directly.

Previously, it was counting the total number of commits in the
repository on the specified branch (could be as high as 20k+), which
requires walking the entire git history to the very first commit. And
comparing the total against requested.

Instead, avoid counting the total number of commits, and simply request
one additional commit (from next page), and check if it's there. If it
is, then HasMore is true.

In my local testing (on a fast SSD), skipping counting all commits
saves 150~ ms on a repo with 21k~ commits, and saves 5~ ms on a repo
with just 2 commits (just the time to spawn a single git process, etc.).