util/handlerutil: fix a deadlock in AugmentCommits
Originally a data race on the map in AugmentCommits had been reported to me by @john
I hadn't had a chance to look into it yet, and so someone got to it before me. They
actually introduced more locking around the map which fixed the data race but caused
a deadlock because `Do` does not return until the currently running `4` rate-limited
goroutines complete (which is impossible because they need the lock to finish!).

The data race also likely was the cause of this issue when loading changesets:

```
rpc error: code = 2 desc = "parsing time \"\" as \"2006-01-02 15:04:05.000000000 -0700\": cannot parse \"\" as \"2006\""
```

Which I've confirmed by loading a changeset previously affected by this issue. A big
thanks to @martin for running into this and sticking with it to try and figure out
what was causing it.