server/testserver: try to remove SGPATH multiple times after waiting for workers to exit
The Changesets tests are flaky because the call to os.RemoveAll on SGPATH is failing. It
fails because we are concurrently writing new files to the directory (the workers are).
I've tried outright directly killing the server process (`ServerCmd.Process.Kill`) but
it happens regardless. As a solution, we can invoke os.RemoveAll a few times in the event
that it fails. This is more future-proof anyway, and in practice does not occur often.