auth: add DefensiveReuseTokenSource for passing good tokens to subcommands
When passing a token to a subcommand, eg for builds, we need to ensure
that the token is valid for a minimum duration so that it does not expire
before the subprocess exits. Before this, we used an oauth2.ReuseTokenSource
for this, which would refresh a token if it is expired. However, it is
possible that a token is valid for only a few minutes, and it won't be
refreshed before passing on to the subcommand, which will subsequently fail.

DefensiveReuseTokenSource will defensively refresh a token if it is valid
for less than 'defensiveExpiry' duration, which is set to 1 hour. So we can
guarantee that the subcommand will have a token that's valid for 1 hour.

Additionally, this change also adds an additional check to possibly refresh
'soon to expire' tokens before passing them to the subcommand.

All this will improve our build stability by a great deal.