app: fix --app.mirror-repo-bg-update
It is not possible to specify `--app.mirror-repo-bg-update=false` because
go-flags does not allow this. Boolean flags are "opt-in", so change the flag to
instead be `--app.no-mirror-repo-bg-update` leaving the default as true/enabled.

Also fix the inverted logic to register the `repoAutoCloner` event listener and
place it inside a `sgxcli.ServeInit` block so that it only runs once the flags
have been parsed.

For reference see https://github.com/jessevdk/go-flags/issues/80#issuecomment-164083839