changesets: remove the git refs store
This change removes the git refs store and the automigration code from
changesets. There is not likely to be any instances still running an
old enough version to not have auto-migrated already, and even if there
are all they need to do is `src serve` on a 0.8.22 binary before
upgrading.