remove defs list alongside files/dirs on repo tree
This removes the list of ~7 def names that are shown to the right of
files and dirs. In practice, this was not very useful either as an
actual feature or as a way to demonstrate Code Intelligence. It also
slowed down page load (not because it blocked the page itself, but
because it blocked loading of other resources, especially when there
were very many dir entries).

We can soon display the last commit or something similar here.