sgx: create clientCtx after all clientCtxFuncs have been set up
f15c201fa78f462fe0e9ba80c3a294cbc0dfea76 moved the creation of
`clientCtx` above some subsequent additions to `clientCtxFuncs`.
Because `clientCtx` was computed in a IIFE, it was not augmented with
the `sharedSecretToken` that was later added to the `clientCtxFuncs`.

This removes the confusing/error-prone IIFE and moves the `clientCtx`
creation to after all the `clientCtxFuncs` have been added.

This fixes tests that attempted to access the Web app server when run
with `--auth.source=none`. The bug was concealed when
`--auth.source=local` because the user's auth token would shadow the
shared secret token (which is in itself desirable because the user's
auth token is strictly more informative).