sgx: simpler way of mounting "srclib" subcommand
Previously we used a forked go-flags with an AddExistingCommand to
mount the "src srclib" command (which is basically a built-in version
of the srclib CLI).

This uses the srclib `mountable-cmds` branch (see PR at
https://github.com/sourcegraph/srclib/pull/235, which should also be
merged) to expose a more standard way of doing this.

As a plus, the srclib subcommands now work in bash completion for the
src command.

The motivation of this change is to reduce the divergence of our
go-flags fork.