standardize config flags in INI file to match CLI flags
Previously, we would tell users "set the CLI flag --foo-bar or the
/etc/sourcegraph/config.ini [serve.Baz] FooBar property". This is
because the INI file required the use of the "command group" name,
whereas the CLI flags do not. Upshot: our docs were way longer than
necessary because they had to explain 2 confusing, different ways of
achieving the same result (the desired configuration).

This changeset standardizes on using INI config directives that match
the equivalent CLI flags. It also incorporates
https://github.com/jessevdk/go-flags/pull/158, which simplifies
go-flags' INI parsing to allow for this.