grpc: Prevent Fatal when returning `nil, nil`
gRPC does not allow services to return nil, nil and will do a Fatal if it
encounters this. However, we can't enforce this at compile time and more than
once we have accidently done this in production causing downtime. Instead we can
return an error code if we break this contract.