worker: continue after failures on auto-generated "test" CI build steps
This is because the goal of the auto-generated CI build steps is to
prepare the environment for srclib analysis, which requires
build/install/deps to run but not the tests. So, we can get by with
Code Intelligence even if tests fail. Of course, if a repo has a
.drone.yml file that's explicit, then those build sections will not be
marked as allowing failures.