server/local: after opening repo, allow caching it for the ctx lifetime
Previously, each call to RepoVCS.Open on a git repo would require
parsing all of the .git/**/*.idx files, which could take several
milliseconds or more and which created a lot of garbage for the GC to
clean up.

This commit introduces an opt-in mechanism whereby server code can
store an opened repo in a memory cache that lives only for the
lifetime of the ctx. This should have minimal adverse effects because
it is short-lived and opt-in.

This change significantly improves performance of Deltas.ListFiles,
which previously could result in calling RepoVCS.Open thousands of
times (one for each hunk '+'/'-'/' ' region). That is the method that
most needed this improvement.