New local install docs don't use install.sh, instead they have you download a single file
Merge changeset #534

install.sh was flaky for some customers (though it works great in a cloud environment), plus this removes the "what is that crazy script doing" concern.

I've run this locally, and toggled between the various options and it works. I still need to test this on RedHat and Centos