platform/apps/changesets: more closely tie TestChangesets_RebaseFlow to the implementation
By only validating the Base and Head of the DeltaSpec if they are specified, we are
more loosly coupled to the implementation. In an ideal world this would be fine, but
the frontend doesn't correctly render the commit history if the Base is not specified
pre-merge on a rebased + force pushed branch (I've verified this with Quinn's fix
reverted). So, instead, couple this test tightly with the implementation to ensure
that such a regression does not occur again.