worker: Rate limit calls to Builds.DequeueNext
We expect build jobs to take at least a few seconds (more likely in the minutes
range). This change conservatively estimates builds should take at least a
second. One potential issue with rate limiting is if the queue is flooded with
many invalid builds, but I haven't seen that happen in practice. If this does
happen, we will need to implement some sort of draining / bursting.