worker: don't call os.Exit when timeout; handle killed builds better
* Updates drone-exec in Godep to make it take a context and respect
  cancelation and deadlines, instead of managing its own timeouts and
  calling os.Exit (which does not work in our case since we are using
  drone-exec as a library and os.Exit kills our whole program). Fixes
  https://src.sourcegraph.com/sourcegraph/.tracker/253 and actually
  reduces the number of stray Docker containers (which are now killed
  when the worker process is killed, unless it's a SIGKILL, of course).

* Makes the worker mark in-progress builds as killed when the worker
  process terminates (due to ctrl-C/SIGTERM/SIGINT). This means that
  when the server is restarted (e.g., because an update is deployed),
  the builds will not linger in a phantom "active" state until the build
  reaper reaps them; they will be immediately marked as killed (which is
  displayed in the UI with red, same as failed). As a future
  improvement, we could make killed builds automatically
  retried/resumed, but that is not necessary right now.

It is of course impossible to guarantee that builds will be cleaned up
(e.g., if you yank the server's power cord). But those will still be
cleaned up (and marked as failed + killed) in a few minutes by the
build reaper.

Note the distinction between failed and killed. A build killed by a
SIGTERM or ctrl-C is not a failure because it's not the build's fault
that someone hit ctrl-C. But if the build is reaped (because it didn't
heartbeat in the required interval), it is considered failed because
it's likely that the build took down the server. We can change these
semantics later if needed.