server/internal/store: have the pgsql backend properly delete repos
The pgsql backend delegates VCS operations to the fs backend, i.e. it's
using the $SGPATH/repos dir as scratch paper to perform it's git-related
operations. Prior to this change, though, we didn't delegate the deletion
operations to the fs backend, and the directory would stick around forever.

The `TestRepos_CreateDeleteWorks_lg` test is now passing.