changesets: do not use strict slashes
Prior to this change visiting `/.changes/1/files/a/` would attempt to redirect
to `/1/files/a` which is misleading because it would still show you all files
under that path (e.g. `a/b` or `a/b/c` would both be matched by just `/1/files/a`).

Remove the strict-slash option so that we get more clearly defined filter URLs
and also to work around a platform bug in the handling of HTTP redirect status
codes (to be fixed in a separate CS).