builds: Don't clear non-boolean state on Update
The boolean fields are only ever set on the final update, so the current logic
of always setting them to false if not set is fine. Otherwise previously we were
clearing out `host`, `priority` and `builderconfig` if they weren't set in the
update. `builderconfig` is particularly annoying to lose, since it helps with
debugging issues in the build.