Improve edge case behavior of handlerutil.HandlerWithErrorReturn error handling.
The original issue is that it's possible for handlerutil.HandlerWithErrorReturn
to write a blank page with 200 OK status with a blank body. This is
confusing for users and developers, and it's not intended.

handlerutil.HandlerWithErrorReturn is composed of a main Handler,
and Error which is the handler that is called when Handler returns an error.

Unexpected things can happen if Handler itself panics or doesn't properly
handler internal errors.

The solution is to require the Error func of handlerutil.HandlerWithErrorReturn
to never panic. This can be achieved either by writing simple code
that is guaranteed to not fail (barring out of memory or hardware failures),
or if external untrusted code must be called (for example, platform apps),
then use recover() to catch and process any panics.

An alternative solution was to add a 2nd tier error handler for when the
original error handler panics, but that proved to be counter-productive.
Most current error handlers are custom and very simple. Only the main app
error handler named HandleError needed to be modified to add a recover
to catch possible panics when executing a template. That template calls out
to external, untrusted code (platform apps) and therefore may panic.

In HandleError, if the template fails to display the error message, fall back
to using a simple plain text error message instead and try to serve it
with 500 status code. It may not work out if bytes have already been
written to http.ResponseWriter, but we can't do anything about that
unless we buffer all writes to http.ResponseWriter before sending them
to the wire, which we currently don't do (it would have some negative
performance implications that must be weighted carefully).

Use structured logging and improve error message display.