changesets: allow updating changeset descriptions
This change causes a pencil icon to display next to a changeset description which
allows the author of the changeset to update it. There is a minor hack employed
to ensure that after submission of a new description the description does not
flicker as if it was the old one (because the old description would still be in
the changeset, and our MarkdownView goes back to the server in order to render
the markdown itself).