Refactored git tests.
There was a lot of duplicate code in the old tests. Refactored the
tests so that there are basically two parameters for a test: a server
config and a set of expectations for various operations that will run
against the server.

On Circle CI the requests being made with OAuth tokens inside URLs
where hitting the 255 char limit of old versions of curl. So I updated
curl and git on Circle CI.

Still need to add tests for pushing over HTTP/SSH with various
authentication scopes.