changesets: enable seeing changed files even with a filter enabled
Prior to this change, you could apply a filter to a changeset via the hidden
feature of appending `/files/dir_or_file` on a changeset URL. It would simply
entirely hide filtered files from the view.

After this change, the files will still appear in the stats view (`Files (3)`
below) and will still have their diff panel show up but without the actual diff
hunks being shown (not rendered or transfered to browser). A new `View Diff`
button is added on the side of these filtered files which currently brings you
to the same CS but with that filter applied (i.e. for viewing only that file).

In the future, we will reconsider this design / add expansion or something else
entirely.

------------------

![before](http://virtivia.com:27080/1xsu6h1hzrfmv.png)

------------------

------------------

![after](http://virtivia.com:27080/6vwijnvo9opg.png)

------------------