changesets: properly implement the LGTM button
Prior to this change we would not show the LGTM button as checked even if you
previously checked it (after a page reload, for example). Fix this by checking
the reviewers LGTM status against the current user.

Also use `onChange` and not `onChecked` (this has the same effect) or else React
will throw a warning:
```
Warning: Failed form propType: You provided a `checked` prop to a form field without an `onChange` handler. This will render a read-only field. If the field should be mutable use `defaultChecked`. Otherwise, set either `onChange` or `readOnly`. Check the render method of `CodeReviewControlPanel`.
```