Restore DebugMode for admin users.
Fixes https://src.sourcegraph.com/sourcegraph@master/.tracker/304.

This change re-enables DebugMode to be on for logged in users with
Admin access rights.

It does so by fetching the full user object for each HTTP request in
the UserMiddleware (which is potentially expensive, explained below).

It's enabled by previous change to use local users (rather than oauth),
and partially started in CS 619. It partially reverts
https://src.sourcegraph.com/sourcegraph/.commits/197dfa87daed43e44c5b2a1a4775b8c6f2ea82b6.

Add a "full user" value to context for each app HTTP request. It can be
used by interested parties to see if user has Admin access rights.

This approach is flexible and can be adjusted in the future to meet
evolving needs, without affecting a lot of code right now.

Timing the additional cost seems reasonable, and can be revisited when
it becomes a larger bottleneck:

	INFO[02-11|14:29:50] ✱ Sourcegraph running at http://localhost:3080
	fetching user taken: 3.600197ms
	fetching user taken: 17.457µs
	fetching user taken: 20.737µs
	fetching user taken: 12.532µs
	fetching user taken: 13.667µs
	fetching user taken: 20.023µs
	fetching user taken: 11.548µs
	fetching user taken: 2.945271ms
	fetching user taken: 2.702866ms
	fetching user taken: 12.033µs
	fetching user taken: 11.497µs
	fetching user taken: 21.828µs
	fetching user taken: 11.699µs
	fetching user taken: 13.008µs
	fetching user taken: 30.488µs