server: added access checks around calls to graph store
All calls made to the GraphStore are protected by access checks
to verify that the user has access to the requested repo. These
checks are places in the server code that calls into the graph store.
This isn't very secure as future changes may add unprotected calls
to the graph store.

A longer term solution is to create a wrapper for GraphStore in the
store package, which handles auth checks before delegating to the
srclib store methods.