pkg/vcs: add LRU cache for ls-tree, show (file), and diff
Caching these low-level operations yields performance benefits for all
of their callers without needing to use the complex, error-prone
grpccache at the highest level. (At this low git level, we don't need
to care about user authentication or eviction, which is a big win in
terms of cache hit rate.)

The LRU sizes are arbitrary. We can tune as needed.