changesets: add temporary code to perform a full index rebuild
This is needed as we (and potentially some users) have run through the data
migration steps to move onto pgsql, and potentially already invoked a
Storage.Delete operation in changesets (by e.g. changing a CS status to closed
or open) which would effictively delete the entire changeset index.

This code runs once automatically on startup and rebuilds the entire index from
the existing data in postgres, so no additional work is needed. The rebuild
completes in around ~60s for our ~500 changesets, and after running all indexes
are in a good state once again.

No change is required anywhere else, as only Changesets makes use of the platform
storage Delete method currently.