global_refs: Prevent duplicate work for indexing
Currently on every srclib import we trigger a refresh of the index. However,
quite often the import is unrelated to the default branch, so we end up doing
duplicate work. Worse still, we can be indexing the same commit concurrently.
Now we store the last indexed commit + we update that value inside of the
transaction. This means we should avoid concurrent updates.

Note there are still race conditions in this, but with the logging we have if
they trigger we will know. I'd like to get it out and see how it behaves in
practice.

Test Plan: unit tests, inspect logging to see we only updated once for older
test. Add new test to ensure updates work.