app/internal/gitserver: remove duplicative TestGitServerWithAuth
The AllowAllLogins/AllowAnonymousReaders auth flags used to enable
different behaviors that required us to test gitserver both WITH and
WITHOUT auth. Now, there is no concept of "server-wide" readers/writers,
so there is no need to test both cases.

All of the test cases (in var tests) except for the first are identical
to the ones in TestGitServerWithAnonymousReaders. The sole exception
tests that anonymous users may not clone repos, but the current desired
behavior is that anonymous users may clone any repo hosted on the server
itself. (This does not let anonymous users clone mirrored repos, such as
private GitHub repos.)

This failing test was not being picked up in CI due to a mistake in
dev/circle-ci-run-tests.sh, which will be fixed in a subsequent commit.