{api,services}: add repo origin fields for storing repo ID on origin service (e.g., GitHub)
Currently the only way to associate a Sourcegraph mirror repo with its
origin on (e.g.) GitHub is by the repo's URI. This is an unreliable
relationship, as a repo's name can change on GitHub, or it can be
deleted and a new (different) repo can assume the previous name. The
only way to have a stable association is to use the immutable numeric
repo ID.

This change introduces new fields to the Repo protobuf message and repo
DB table to store this association: origin repo ID, origin service type,
and origin API URL. These 3 pieces of information are necessary to
support multiple repo origin hosts. See the protobuf field documentation
for more information. The values in these fields are populated but not
yet used.

There is a manual SQL migration that must be run. In addition, this
change automatically and gradually migrates repo
rows (when (localstore.repos).Get is called on them) in the DB to
contain the correct origin values. This migration code can be removed
when the DB mirror repos all have values set for these fields.

Some repos in the DB might be in conflict; e.g., there may be 2 rows for
the same repo, one with its old name and one with its new name in the
URI column. In this case, we'll need to manually delete the repo with
the old name.

Migration SQL:

  begin;
  alter table repo add column origin_repo_id text;
  alter table repo add column origin_service integer;
  alter table repo add column origin_api_base_url text;
  commit;

After all repos have been migrated and duplicates have been removed,
you can run the following migration SQL. It will be mentioned again in
a future commit message after the migration is complete.

  create unique index repo_origin on repo(origin_repo_id, origin_service, origin_api_base_url) where origin_repo_id is not null;