services/ext/slack: better default channel handling
Prior to this change we had to either pass default channel values for
Slack notifications by explicitly hard-coding a value, passing it via
the environment, or using the default `dev-bot` channel set in the CLI
config.

When we send a channel name to the Slack API, _it acts as an override
for the already default channel name for the given webhook_. That is,
if we don't provide a default channel name then we can configure where
notifications go per webhook URL via Slack's web UI which is more
accessible.